base mode complete

+++ server/public/client.js
@@ -14,19 +14,10 @@ function getTodoList() {
     }).then((response) => {
         $('#todo-list').empty();
         response.forEach(task => {
-            if (task.is_complete === true) {
+            
+            if (task.priority === 'Urgent, important') {
                 $('#todo-list').append(`
-                    <tr class"task-complete">
-                        <td>${task.task}</td>
-                        <td>${task.priority}</td>
-                        <td></td>
-                        <td><button class="delete-button" data-id="${task.id}">Delete</button></td>
-                    </tr>
-                `);
-            }
-            else {
-                $('#todo-list').append(`
-                    <tr>
+                    <tr class="row-ui"> 
                         <td>${task.task}</td>
                         <td>${task.priority}</td>
                         <td><button class="complete-button" data-id="${task.id}">Complete</button></td>
@@ -34,6 +25,48 @@ function getTodoList() {
                     </tr>
                 `);
             }
+            else if (task.priority === 'Not urgent, important') {
+                $('#todo-list').append(`
+                <tr class="row-nui"> 
+                    <td>${task.task}</td>
+                    <td>${task.priority}</td>
+                    <td><button class="complete-button" data-id="${task.id}">Complete</button></td>
+                    <td><button class="delete-button" data-id="${task.id}">Delete</button></td>
+                </tr>
+                `);
+            }
+            else if (task.priority === 'Urgent, not important') {
+                $('#todo-list').append(`
+                <tr class="row-uni"> 
+                    <td>${task.task}</td>
+                    <td>${task.priority}</td>
+                    <td><button class="complete-button" data-id="${task.id}">Complete</button></td>
+                    <td><button class="delete-button" data-id="${task.id}">Delete</button></td>
+                </tr>
+                `);
+            }
+            else if (task.priority === 'Not urgent, not important') {
+                $('#todo-list').append(`
+                <tr class="row-nuni"> 
+                    <td>${task.task}</td>
+                    <td>${task.priority}</td>
+                    <td><button class="complete-button" data-id="${task.id}">Complete</button></td>
+                    <td><button class="delete-button" data-id="${task.id}">Delete</button></td>
+                </tr>
+                `);
+            }
+
+            if (task.is_complete === true) {
+                $('#todo-list').append(`
+                    <tr class="row-complete">
+                        <td>${task.task}</td>
+                        <td>${task.priority}</td>
+                        <td></td>
+                        <td><button class="delete-button" data-id="${task.id}">Delete</button></td>
+                    </tr>
+                `);
+            }
+
         });
     });
 }
@@ -65,11 +98,14 @@ function handleClickCompleteTask() {
 }
 
 function handleClickDeleteTask() {
-    const id = ($(this).data().id);
-    $.ajax({
-        method: 'DELETE',
-        url: `/tasks/${id}`
-    }).then(() => {
-        getTodoList();
-    });
+    const result = confirm(" Are you sure you want to delete this task?")
+    if (result) {
+        const id = ($(this).data().id);
+        $.ajax({
+            method: 'DELETE',
+            url: `/tasks/${id}`
+        }).then(() => {
+            getTodoList();
+        });
+    }
 }
